+++ csatar-plugins/knowledgerepository/lang/en/lang.php
@@ -28,6 +28,10 @@
                     'location' => 'Location',
                     'gameTypes' => 'Game Types',
                     'gameType' => 'Game Type',
+                    'methodologyType' => 'Methodology type',
+                    'methodologyTypes' => 'Methodology types',
+                    'methodologyName' => 'Methodology name',
+                    'ageGroup' => 'Age group',
                 ],
             ],
             'general' => [
@@ -39,6 +43,11 @@
                 'proposerCsatarCode' => 'Proposer',
                 'isApproved' => 'Approved',
                 'minute' => 'minute',
+                'link' => 'Link',
+                'Attachment' => 'Attachment',
+                'sortOrder' => 'Sort order',
+                'version' => 'Version',
+                'created_at' => 'Created at'
             ],
         ],
     ],