+++ csatar-plugins/knowledgerepository/updates/builder_table_update_csatar_knowledgerepository_methodologies.php
@@ -1,0 +1,27 @@
+<?php namespace Csatar\KnowledgeRepository\Updates;
+
+use Schema;
+use October\Rain\Database\Updates\Migration;
+
+class BuilderTableUpdateCsatarKnowledgerepositoryMethodologies extends Migration
+{
+    public function up()
+    {
+        Schema::table('csatar_knowledgerepository_methodologies', function($table)
+        {
+            $table->timestamp('deleted_at')->nullable();
+            $table->integer('sort_order')->nullable();
+            $table->dateTime('version')->nullable();
+        });
+    }
+    
+    public function down()
+    {
+        Schema::table('csatar_knowledgerepository_methodologies', function($table)
+        {
+            $table->dropColumn('deleted_at');
+            $table->dropColumn('sort_order');
+            $table->dropColumn('version');
+        });
+    }
+}